web/requirements.txt
attrs
dateparser
eventlet
flask
flask-socketio
redis
requests
